backend/fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.36
PyMySQL==1.1.1
cryptography==44.0.3
python-dotenv==1.0.1
pydantic==2.10.6
passlib[bcrypt]==1.7.4
xlrd==2.0.1
